+++ drasyl-cli/src/main/java/org/drasyl/cli/perf/PerfServerCommand.java
@@ -69,7 +69,7 @@ public class PerfServerCommand extends ChannelOptions {
     @Override
     protected ChannelHandler getHandler(final Worm<Integer> exitCode,
                                         final Identity identity) {
-        return new PerfServerChannelInitializer(identity, bindAddress, networkId, onlineTimeoutMillis, superPeers, out, err, exitCode);
+        return new PerfServerChannelInitializer(identity, bindAddress, networkId, onlineTimeoutMillis, superPeers, out, err, exitCode, !protocolArmDisabled);
     }
 
     @Override