+++ drasyl-cli/src/main/java/org/drasyl/cli/wormhole/WormholeReceiveCommand.java
@@ -90,7 +90,7 @@ public class WormholeReceiveCommand extends ChannelOptions {
     @Override
     protected ChannelHandler getHandler(final Worm<Integer> exitCode,
                                         final Identity identity) {
-        return new WormholeReceiveChannelInitializer(identity, bindAddress, networkId, onlineTimeoutMillis, superPeers, err, exitCode, code.first());
+        return new WormholeReceiveChannelInitializer(identity, bindAddress, networkId, onlineTimeoutMillis, superPeers, err, exitCode, code.first(), !protocolArmDisabled);
     }
 
     @Override